Write a function for user's and computer's input
Add a function that asks the user for input, allowing for choice to be
entered. The function also gets the computer's random choice and compares
the user's and choice with the computer's choice. Depending on the
inputs of the two sides, a pop-up message about the result gets printed
on the screen as a result.

The user's and computer's inputs were made case insensitive by adding
the .toLowerCase() to both inputs so that when if statements are made
to compare the two results, only one variation of the inputs is needed.